+++ test/document/revisions_tests.dart
@@ -68,7 +68,11 @@ class RevisionsTests
       documentData
     );
 
-    await context.getApi().acceptAllRevisionsOnline(request);
+    final result = await context.getApi().acceptAllRevisionsOnline(request);
+    expect(result.document, isNotNull);
+    expect(result.model, isNotNull);
+    expect(result.model.result, isNotNull);
+    expect(result.model.result.dest, isNotNull);
   }
 
   /// Test for rejecting revisions in document.
@@ -97,6 +101,10 @@ class RevisionsTests
       documentData
     );
 
-    await context.getApi().rejectAllRevisionsOnline(request);
+    final result = await context.getApi().rejectAllRevisionsOnline(request);
+    expect(result.document, isNotNull);
+    expect(result.model, isNotNull);
+    expect(result.model.result, isNotNull);
+    expect(result.model.result.dest, isNotNull);
   }
 }